Add setting extensions to get setting value
Add a `key` property to the `Setting` collection which will correspond
to a dictionary key which can be used to get the setting value.

Add a `get` method on `Setting` which can be used to get the value of a
particular key. If there isnt' a setting with that key, `undefined` will
be returned. If there's no value for that key, the `defaultValue` will
be used.